requirements.txt
google-generativeai>=0.3.0
pyyaml>=6.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
